refactor(network-proxy): rename attempt metadata module
