+++ codex-rs/network-proxy/src/http_proxy.rs
@@ -1,5 +1,5 @@
-use crate::attempt_metadata::attempt_id_from_proxy_authorization;
 use crate::config::NetworkMode;
+use crate::metadata::attempt_id_from_proxy_authorization;
 use crate::network_policy::NetworkDecision;
 use crate::network_policy::NetworkDecisionSource;
 use crate::network_policy::NetworkPolicyDecider;
