+++ codex-rs/network-proxy/src/lib.rs
@@ -1,9 +1,9 @@
 #![deny(clippy::print_stdout, clippy::print_stderr)]
 
 mod admin;
-mod attempt_metadata;
 mod config;
 mod http_proxy;
+mod metadata;
 mod network_policy;
 mod policy;
 mod proxy;
