+++ codex-rs/network-proxy/src/metadata.rs
